# app/api/ai/research/route.ts
/**
 * Research Agent API Endpoint
 *
 * POST /api/ai/research
 *
 * Generates market research and audience insights based on brand brief.
 * Includes authentication, budget enforcement, and full usage tracking.
 */

import { NextRequest, NextResponse } from 'next/server';
import { cookies } from 'next/headers';
import { requireProjectOwnership, handleAuthError } from '@/lib/auth/middleware';
import { checkBudgets, logLLMUsage, saveLLMOutput } from '@/lib/auth/helpers';

/**
 * Request body interface
 */
interface ResearchRequest {
  projectId: string;
  brief: string;
  options?: {
    model?: string;
    provider?: string;
    temperature?: number;
    maxTokens?: number;
  };
}

/**
 * Response interface
 */
interface ResearchResponse {
  success: boolean;
  data?: {
    research: string;
    metadata: {
      model: string;
      provider: string;
      tokensUsed: number;
      cost: number;
      latencyMs: number;
      cached: boolean;
    };
  };
  error?: {
    code: string;
    message: string;
  };
}

/**
 * POST /api/ai/research
 *
 * Generate brand research from a brief.
 */
export async function POST(request: NextRequest) {
  try {
    // 1. Parse and validate request body
    const body: ResearchRequest = await request.json();

    if (!body.projectId || !body.brief) {
      return NextResponse.json(
        {
          success: false,
          error: {
            code: 'INVALID_REQUEST',
            message: 'projectId and brief are required',
          },
        },
        { status: 400 }
      );
    }

    // 2. Authenticate and verify project ownership
    const { user, project, error: authError } = await requireProjectOwnership(body.projectId);
    if (authError) return authError;

    // 3. Check budget limits
    const budgetCheck = await checkBudgets(body.projectId, user.id);
    if (!budgetCheck.allowed) {
      return NextResponse.json(
        {
          success: false,
          error: {
            code: 'BUDGET_EXCEEDED',
            message: budgetCheck.message || 'Budget exceeded',
          },
        },
        { status: 429 }
      );
    }

    // TODO: 5. Get LLM Provider Manager
    // import { getLLMProviderManager } from '@/lib/ai/manager';
    // const manager = getLLMProviderManager();

    // TODO: 6. Load research prompt template
    // import { createResearchPrompt } from '@/lib/ai/prompts/research';
    // const prompt = createResearchPrompt({
    //   brief: body.brief,
    //   industry: project.data.industry,
    //   targetAudience: project.data.target_audience,
    // });

    // TODO: 7. Generate research
    // const result = await manager.generateCompletion({
    //   prompt,
    //   systemPrompt: 'You are a brand research specialist...',
    //   taskType: 'research',
    //   model: body.options?.model,
    //   provider: body.options?.provider,
    //   temperature: body.options?.temperature ?? 0.7,
    //   maxTokens: body.options?.maxTokens ?? 4096,
    //   metadata: {
    //     projectId: body.projectId,
    //     userId: session.user.id,
    //     taskType: 'research',
    //   },
    // });

    // TODO: 8. Save research to database
    // await supabase.from('research_outputs').insert({
    //   project_id: body.projectId,
    //   content: result.content,
    //   model: result.model,
    //   provider: result.providerId,
    //   tokens_used: result.usage.totalTokens,
    //   cost: result.cost,
    //   latency_ms: result.latencyMs,
    //   created_by: session.user.id,
    // });

    // TODO: 9. Log usage for analytics
    // await supabase.from('llm_usage_logs').insert({
    //   user_id: session.user.id,
    //   project_id: body.projectId,
    //   provider: result.providerId,
    //   model: result.model,
    //   task_type: 'research',
    //   prompt_tokens: result.usage.promptTokens,
    //   completion_tokens: result.usage.completionTokens,
    //   total_tokens: result.usage.totalTokens,
    //   cost: result.cost,
    //   latency_ms: result.latencyMs,
    //   cached: result.cached,
    //   success: true,
    // });

    // TODO: 10. Return success response
    return NextResponse.json({
      success: true,
      data: {
        research: 'TODO: Implement research generation',
        metadata: {
          model: 'grok-beta',
          provider: 'grok',
          tokensUsed: 0,
          cost: 0,
          latencyMs: 0,
          cached: false,
        },
      },
    });

  } catch (error) {
    console.error('Research generation error:', error);
    return handleAuthError(error);
  }
}

/**
 * GET /api/ai/research
 *
 * Get previous research for a project.
 */
export async function GET(request: NextRequest): Promise<NextResponse> {
  try {
    const { searchParams } = new URL(request.url);
    const projectId = searchParams.get('projectId');

    if (!projectId) {
      return NextResponse.json(
        { success: false, error: { code: 'INVALID_REQUEST', message: 'projectId is required' } },
        { status: 400 }
      );
    }

    // Authenticate and verify project ownership
    const { user, error: authError } = await requireProjectOwnership(projectId);
    if (authError) return authError;

    // TODO: Fetch research outputs from database
    // const supabase = createClient(cookies());
    // const { data: research } = await supabase
    //   .from('outputs')
    //   .select('*')
    //   .eq('project_id', projectId)
    //   .eq('section', 'research')
    //   .order('created_at', { ascending: false })
    //   .limit(10);

    return NextResponse.json({
      success: true,
      data: {
        research: [],
      },
    });

  } catch (error) {
    console.error('Research fetch error:', error);
    return handleAuthError(error);
  }
}