# docs/api-spec.yaml
openapi: 3.0.0
info:
  title: AI Provider API
  version: 1.0.0
  description: |
    Comprehensive API for the Strategist Agent Platform providing AI-powered research, strategy, and critique capabilities with automatic provider fallback, cost tracking, and quality assurance.

    ## Features
    - Multi-provider support (Grok, Claude, OpenAI)
    - Automatic fallback chain for resilience
    - Real-time streaming responses
    - Cost tracking and budget enforcement
    - Response caching for optimization
    - Quality scoring and validation
    - Analytics and usage reporting

  contact:
    name: API Support
    email: support@strategistagent.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.strategistagent.com
    description: Production server
  - url: http://localhost:3000
    description: Local development server
...
security:
  - BearerAuth: []
paths:
  /api/ai/research:
    post:
      tags:
        - AI Generation
      summary: Generate research analysis
      description: |
        Analyzes a creative brief and generates comprehensive research including market insights, competitor analysis, and audience research.

        This endpoint uses the configured AI provider (with automatic fallback) to generate high-quality research outputs.
      operationId: generateResearch
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - projectId
                - brief
              properties:
                projectId:
                  type: string
                  format: uuid
                  description: Unique identifier for the project
                  example: "550e8400-e29b-41d4-a716-446655440000"
                brief:
                  type: string
                  description: Creative brief content to analyze
                  example: "Launch campaign for eco-friendly sneakers targeting Gen Z consumers..."
                  minLength: 50
                  maxLength: 50000
            example:
              projectId: "550e8400-e29b-41d4-a716-446655440000"
              brief: "Launch campaign for eco-friendly sneakers targeting Gen Z consumers. Brand values: sustainability, authenticity, innovation."
      responses:
        '200':
          description: Research generated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResearchResponse'
              example:
                success: true
                outputId: "660e8400-e29b-41d4-a716-446655440001"
                content: |
                  # Market Research Analysis

                  ## Market Insights
                  - Sustainable fashion market growing at 9.7% CAGR
                  - Gen Z willing to pay 10% premium for eco-friendly products

                  ## Competitor Analysis
                  1. Allbirds: Leading sustainable footwear brand...
                  2. Veja: French eco-sneaker brand...

                  ## Target Audience
                  - Age: 18-25
                  - Values: Environmental consciousness, social justice
                metadata:
                  provider: "grok"
                  model: "grok-2-latest"
                  cost: 0.0234
                  cached: false
                  tokensUsed: 4680
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: Project not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: "Project not found"
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /api/ai/critique:
    post:
      tags:
        - AI Generation
      summary: Generate quality critique
      description: |
        Analyzes and critiques generated strategy, providing quality assessment, improvement suggestions, and validation.
      operationId: generateCritique
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - projectId
                - strategyData
              properties:
                projectId:
                  type: string
                  format: uuid
                strategyData:
                  type: string
                  description: Strategy output to critique
            example:
              projectId: "550e8400-e29b-41d4-a716-446655440000"
              strategyData: "# Brand Strategy\n\n## Positioning..."
      responses:
        '200':
          description: Critique generated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResearchResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Supabase authentication token. Include in Authorization header as:
        `Authorization: Bearer YOUR_SUPABASE_TOKEN`
  schemas:
    ResearchResponse:
      type: object
      required:
        - success
        - outputId
        - content
        - metadata
      properties:
        success:
          type: boolean
          description: Indicates if the operation was successful
          example: true
        outputId:
          type: string
          format: uuid
          description: Unique identifier for the generated output
        content:
          type: string
          description: Generated content (Markdown format)
        metadata:
          type: object
          required:
            - provider
            - model
            - cost
            - cached
            - tokensUsed
          properties:
            provider:
              type: string
              description: LLM provider used
              enum: [grok, claude, openai]
            model:
              type: string
              description: Specific model used
              example: "grok-2-latest"
            cost:
              type: number
              format: double
              description: Cost in USD
              example: 0.0234
            cached:
              type: boolean
              description: Whether response was served from cache
              example: false
            tokensUsed:
              type: integer
              description: Total tokens consumed
              example: 4680
...
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Error message
        details:
          type: string
          description: Additional error details (only in development)
        totalCost:
          type: number
          description: Current total cost (for budget errors)
        userTotal:
          type: number
          description: User's total cost (for budget errors)
  responses:
    BadRequest:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "projectId and brief are required"
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Unauthorized"
    Forbidden:
      description: Access forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Forbidden"
    RateLimitExceeded:
      description: Budget or rate limit exceeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            projectBudget:
              summary: Project budget exceeded
              value:
                error: "Project budget exceeded. Max: $10"
                totalCost: 10.42
            userBudget:
              summary: User monthly budget exceeded
              value:
                error: "Monthly budget exceeded. Max: $100"
                userTotal: 103.56
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Internal server error"
            details: "Stack trace available in development mode"